omaigaaaa empezando con streaming--> recien empezando con websocket de parte de cliente

+++ lib/features/streaming/dto/comando.dart
@@ -1,7 +1,9 @@
 
+import 'dart:ffi';
+
 class ComandoDTO {
   final String comando;
-  final int segundos;
+  final Long segundos;
   final String musicId;
 
   ComandoDTO({
